+++ guillotina/tests/fixtures.py
@@ -4,6 +4,7 @@ from unittest import mock
 
 import aiohttp
 import pytest
+import json
 from aiohttp.client_exceptions import ContentTypeError
 from aiohttp.test_utils import TestServer
 from guillotina import testing
@@ -424,3 +425,17 @@ def redis_container():
     yield host, port  # provide the fixture value
     pytest_docker_fixtures.redis_image.stop()
     annotations['redis'] = None
+
+
+class DBUsersRequester(ContainerRequesterAsyncContextManager):
+    async def __aenter__(self):
+        requester = await super().__aenter__()
+        await requester(
+            "POST", "/db/guillotina/@addons", data=json.dumps({"id": "dbusers"})
+        )
+        return requester
+
+
+@pytest.fixture(scope="function")
+async def dbusers_requester(guillotina):
+    return DBUsersRequester(guillotina)